+++ src/Slots/SuperheroSlotMachine.java
@@ -4,7 +4,7 @@ package Slots;
 public class SuperheroSlotMachine extends SlotMachinesTemplate {
     private static final String[] Superheroes = {"Batman", "Superman", "Ironman"};
 
-    public SuperheroSlotMachine(double initialBalance) {
+    public SuperheroSlotMachine(int initialBalance) {
 
         super(initialBalance);
     }
